Allow SubCommand to have a flag and arguments
This commit refactors the subcommand flag handling to allow having flags
and arguments at once. For additional arguments, the existing MinArgs
and MaxArgs options are handled.

Furthermore, when generating a '--help' page, this commit adds an
additional 'FlagDescription' script. This is used to generate the help
text for the flag option only.