Add 'godoc' and 'localgodoc' ggman web base urls
This commit adds two new predefined base urls for ggman web and ggman
url. They are 'godoc' and 'localgodoc' and point to 'https://pkg.go.dev'
and 'http://localhost:6060/pkg/' respectively. Unlike previous builtins,
these urls also appends the base url to the url in question. To achieve
this, a new flag was added to WebBuiltInBases.