Add 'find-branch' command
git 2.28 introduced the `init.defaultBranch` option to set the name of
the default branch of new repositories. However this does not affect
existing repositories.

To find repositories with an old branch, this commit introduces the
`find-branch` command. It takes a single argument (a branch name), and
finds all repositories that contain a branch with the given name.