Add util.NullStream to ggman.IOStream methods
This commit adds a new utility, util.NullStream, that acts as a no-op
implementation of io.Reader and io.Writer.

It then updates the NewIOStream method to assume sensible defaults when
nil is passed to it. It furthermore adds a convenience method that
always returns an IOStream with the defaults.