Only create parent folder when cloning repository
Previously, when cloning a repository, the folder the repository would
be cloned into would be created automatically. This resulted in a lot of
unuused folders when a typo was made.

In order to combat this, this commit changes the clone behavior. Only
the parent folder the repository will be cloned into would be created.