env.ScanRepos: Use Filter Candidates properly
In a previous commit new code to use Candidates of a filter during the
env.ScanRepos was detected. An analysis of the source code turned out
that this filtering was not actually effective because of a missing
statement. This was not detected earlier because so far no use case has
required this filtering.

This commit fixes the issue by adding the missing statement.