util/scanner: Rename Filter -> Visit
This commit renames the Filter function to Visit, in order to clarify
the semantics of the scanner. Instead of filtering a list of directories
it instead visits every directory. The collection of results is a
side-effect of this visitation.

On top of this, the new visit function now receives a context argument
that indicates the context in which the directory was visited.