env.Matches: Use relative path for empty remote
Currently when using a '--for' filter (via the env.Matches function) and
a repository is missing a remote, it would always match. This was not
the intended behavior, as odd repositories would show up in certain
matches.

Instead this commit updates the function to use the relative path to the
root directory.