Add force-repo-here flag to ggman web & ggman url
This commit adds a new flag called 'force-repo-here' to the 'ggman web'
and 'ggman url' commands. It allows pretending there is a repository at
the current working directory even when there is none.

This is intended to open the web browser at a page corresponding to a
remote namespace e.g. 'github.com/tkw1536'.