env/url.go: Add 'URL.IsLocal()' method
Previously, the 'ggman clone' command was incorrectly reporting some
urls as local, even when they were not. This was caused by an incorrect
check if a URL is local.

This commit adds a new method `url.IsLocal()' that that replaces the old
behavior. It properly checks if a URL looks like a local url.