Normalize Local repository paths
Previously cloning a repository used the exact components of the remote
URI. When one accidentally cloned a directory with a different casing
than intended this caused unexpected directory clutter on case-sensitive
file systems.

This commit updates the ranslation of URLs into paths to take existing
paths into account. In particular, it re-uses existing sub-paths if they
differ from the requested path only by casing.

By default, the first matching directory (in alphanumerical order) is
used as opposed to creating a new one. If a directory with the exact
name exists, this is prefered over a case-insensitive match.

This normalization behavior can be controlled using a new `GGNORM`
environment variable.