internal/testutil: Refactor TempDir => TempDirAbs
This commit refactors and renames the TempDir function in the
internal/testutil package.

Previously this function did not take any parameters and returned a
cleanup function.

The new implementation makes use of the TempDir method of the testing.T
package. This requires passing the testing.T instance and absolves the
need for having an explict cleanup.

Unlike the previous implementation this does not resolve symlinks,
therefore this commit renames TempDir to TempDirAbs to indicate the new
functionality.

Finally the passing of a testing.T requires changes to related functions.