Add 'ggman sweep' command
After moving repositories around (for example using `ggman relocate`, or
by manual operations) empty directories are often left behind. Because
of the nature of ggman sometimes directories containing only empty
directories may also be left behind.

To easily find these directories, this commit introduces the `ggman
sweep` command. It takes no arguments, and lists all directories, which
are not git repositories and are empty, or contain only empty
directories. These are listed in such an order that they can be deleted
in order using `rmdir` and friends.

The ggman sweep command is based on an appropriate refactoring of the
`internal/scanner` package which this commit reworks into an
'internal/walker' package.