env/filter: Add '--from-file' flag
This commit adds a new '--from-file' flag to ggman. This functions
exactly as if one would provide a path via a '--for' argument except
that it is read from a file. Each filter should be on it's own line.
Lines are trimmed for whitespace, and blank lines or lines starting with
';', '//' or '#' are ignored.