{clone,pull,fetch}: Output progress on stderr
This commit updates the 'gogit' implementation to output progress on the
standard error stream as opposed to being mixed into the standard output
stream.

This test furthermore refactors the appropriate test cases to be able to
ignore this output as the exact details are considered an implementation
detail and should not cause testcases to fail.