Introduce `ggman shellrc` command
Previously, the README defined several shell aliases that were useful in
conjunction with ggman. These had to be added to a shell profile
manually.

This commit adds a new `ggman shellrc` command that prints these aliases
to standard output, so that they can be added in a more automatic way.