pattern/SplitPattern: Account for position
This commit reworks the way SplitPatterns compute scores. Previously,
the scores of the sub-patterns were simply added. However, this did not
take into account where in the given parts the pattern was matched.

Consider for example the pattern `hello/world` to be matched against
`i-have-three-levels/hello/world/stuff` and
`i-have-two-levels/hello/world`. Previously these were score
identically.

We now ensure that the latter would receive a higher score, as it more
closely resembles the requested pattern.