Add `--exact-url` flag to `ggman clone`
As noted in #2, sometimes it becomes necessary to clone a repository
using a non-canonical URL on a one-off basis. While it is possible to
configure the CANFILE appropriately, this becomes annoying when done for
individual repositories.

This commit adds a new flag `--exact-url` to `ggman clone` which
simplifies this. When passed, it clones a repository using the exact URL
provided, skipping any canonicalization.