Add GetUsedRemotes and DeleteRemote() to plumbing
This commit adds the two new functions to the plumbing interface.
GetUsedRemotes returns the names of all remotes in used and DeleteRemote
deletes a remote.

It also adds appropriate new function 'CreateTrackingBranch' to the
testutil package to aid in testing. Finally, it removes the file
argument from CommitTestFiles, opting to always auto-generate test file
contents.