cmd/web.go: Properly select current remote
Previously, `ggman url` and `ggman web` without an explicit `--remote`
flag were supposed to use the remote of the current branch. This
functionality was not documented anywhere, untested and did not work as
designed.

This commit fixes the functionality, documents it inside the command
arguments, and adds a test for it.