cmd/relocate: Handle multiple remotes
Previously, the `ggman relocate` command would only consider the current
remote when checking if a repository is in the correct location.

This commit updates the behavior to instead consider a repository
correctly located if it is at the canonical location for any of the
existing remotes. It also adds a new `--only-current-remote` flag to the
command to restore the old behavior.